% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{buildClusters}
\alias{buildClusters}
\title{buildClusters}
\usage{
buildClusters(X, cluster.ids, clust.summary = "mean")
}
\arguments{
\item{X}{[n_sample x n_features] feature matrix}

\item{cluster.ids}{cluster ids returned by function learnClusters (entire list returned)}

\item{clust.summary}{function used to summarize clusters (default "mean")}
}
\description{
This function "builds" clusters involved in the ClusterLasso method, i.e., summarizes features of each cluster into a single variable
}
\examples{
buildClusters()
}
