% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{clusterlasso_cv}
\alias{clusterlasso_cv}
\title{clusterlasso_cv}
\usage{
clusterlasso_cv(
  X,
  y,
  subgroup = NULL,
  n.lambda = 200,
  n.folds = 10,
  n.repeat = 3,
  seed = 27,
  verbose = TRUE,
  screen.thresh = 0.9,
  clust.thresh = 0.9,
  clust.summary = "mean"
)
}
\arguments{
\item{X}{[n_sample x n_features] feature matrix}

\item{y}{[n_sample] vector containing reference labels (+1/-1)}

\item{subgroup}{[n_sample] vector containing a "subgroup" information (e.g., bacterial lineages), to be stratified among the folds (optional)}

\item{n.lambda}{number of regularization values considered (default 200)}

\item{n.folds}{number of folds (default 10)}

\item{n.repeat}{number of repetitions of the cv process (default 3)}

\item{seed}{random seed (default 27)}

\item{verbose}{show progress (default TRUE)}

\item{screen.thresh}{correlation threshold considered to screen variables (default 0.9)}

\item{clust.thresh}{correlation threshold considered to cluster variables (default 0.9)}

\item{clust.summary}{function used to summarize clusters (default "mean")}
}
\description{
This function carries out ClusterLasso-penalized logistic regression cross-validation using glmnet
}
\examples{
clusterlasso_cv()
}
