% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{extract_best_model}
\alias{extract_best_model}
\title{extract_best_model}
\usage{
extract_best_model(
  cv.res,
  modsel.criterion = "balanced.accuracy.best",
  best.eps = 1,
  method = "lasso"
)
}
\arguments{
\item{cv.res}{list containing cross-validation results (produced by lasso_cv() or clusterlasso_cv() )}

\item{modsel.criterion}{prediction performance criterion to select best model (balanced.accuracy.best or auc) (defaut = balanced.accuracy.best)}

\item{best.eps}{AUC tolerance considered to flag best model (defaut = 1)}

\item{method}{string defining the method considered ("lasso" or "clusterlasso" - default = "lasso")}
}
\description{
This function extract the best lasso model from cross-validation results
}
\examples{
extract_best_model()
}
