% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_perf_best_cutoff}
\alias{get_perf_best_cutoff}
\title{get_perf_best_cutoff}
\usage{
get_perf_best_cutoff(cv.res)
}
\arguments{
\item{cv.res}{cross-validation results (list returned by lasso or cluster-lasso cross-validation fuctions)}
}
\description{
This function extract the predictive performance (sensi/speci/accuracy/macro-accuracy) at optimal cutpoints on ROC curves obtained by cross-validation.
}
\examples{
get_perf_best_cutoff()
}
