% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lasso.R
\name{lasso_cv}
\alias{lasso_cv}
\title{lasso_cv}
\usage{
lasso_cv(
  X,
  y,
  subgroup = NULL,
  n.lambda = 200,
  n.folds = 10,
  n.repeat = 3,
  seed = 27,
  verbose = TRUE,
  alpha = 1
)
}
\arguments{
\item{X}{[n_sample x n_features] feature matrix}

\item{y}{[n_sample] vector containing reference labels (+1/-1)}

\item{subgroup}{[n_sample] vector containing a "subgroup" information (e.g., bacterial lineages), to be stratified among the folds (optional)}

\item{n.lambda}{number of regularization values considered (default 200)}

\item{n.folds}{number of folds (default 10)}

\item{n.repeat}{number of repetitions of the cv process (default 3)}

\item{seed}{random seed (default 27)}

\item{verbose}{show progress (default TRUE)}

\item{alpha}{elasticnet parameter (default alpha=1, i.e., lasso)}
}
\description{
This function carries out Lasso-penalized logistic regression cross-validation using glmnet
}
\examples{
lasso_cv()
}
