% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{learnClusters}
\alias{learnClusters}
\title{learnClusters}
\usage{
learnClusters(
  X,
  y,
  screen.thresh = 0.9,
  clust.thresh = 0.9,
  lambda.grid = NULL,
  n.lambda = 200,
  max.size = 250000,
  chunk.size = 10000,
  verbose = TRUE
)
}
\arguments{
\item{X}{[n_sample x n_features] feature matrix}

\item{y}{[n_sample] vector containing reference labels (+1/-1)}

\item{screen.thresh}{correlation threshold considered to screen variables (default 0.9)}

\item{clust.thresh}{correlation threshold considered to cluster variables (default 0.9)}

\item{lambda.grid}{use specified grid of regularization values (default NULL : computed from data)}

\item{n.lambda}{number of regularization values considered (default 200)}

\item{max.size}{maximum size to consider to build entire Nactive x Nfeature correlation matrix (default 250000)}

\item{chunk.size}{chunk size to consider when Nactive x Nfeature correlation matrix is computed by chunk (default 10000)}

\item{verbose}{show progress (default TRUE)}
}
\description{
This function identifies cluster involved in the ClusterLasso method
}
\examples{
learnClusters()
}
