% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{predict_clustlasso}
\alias{predict_clustlasso}
\title{predict_clusterlasso}
\usage{
predict_clustlasso(X, model, method = "lasso", clust.summary = "mean")
}
\arguments{
\item{X}{n.samples x n.features matrix}

\item{model}{(best) lasso / clusterlasso model}

\item{method}{string defining the method considered ("lasso" or "clusterlasso" - default = "lasso")}

\item{clust.summary}{function used to summarize clusters in clusterlasso (default "mean")}
}
\description{
This function makes predictions from the model selected by the extract_best_model function
}
\examples{
predict_clusterlasso()
}
