% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visu.R
\name{show_cv_best}
\alias{show_cv_best}
\title{show_cv_best}
\usage{
show_cv_best(
  cv.res,
  plot.id = "cv-results",
  modsel.criterion = "balanced.accuracy.best",
  best.eps = 1,
  method = "lasso"
)
}
\arguments{
\item{cv.res}{list containing cross-validation results (produced by lasso_cv() or clusterlasso_cv() )}

\item{plot.id}{string containing a label to consider for the plots (typically, the drug corresponding to the cv.res object)}

\item{modsel.criterion}{prediction performance criterion to select best model (balanced.accuracy.best or auc) (defaut = balanced.accuracy.best)}

\item{best.eps}{best-macro-accuracy tolerance considered to flag best model (default = 1)}

\item{method}{string defining the method considered ("lasso" or "clusterlasso" - default = "lasso")}
}
\description{
This function generates a figure (made of 4 individual plots) summarizing the best model identified by cross-validation
}
\examples{
show_cv_best()
}
