% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visu.R
\name{show_cv_overall}
\alias{show_cv_overall}
\title{show_cv_overall}
\usage{
show_cv_overall(
  cv.res,
  plot.id = "cv-results",
  modsel.criterion = "balanced.accuracy.best",
  best.eps = 1
)
}
\arguments{
\item{cv.res}{list containing cross-validation results (produced by lasso_cv() or clusterlasso_cv() )}

\item{plot.id}{string containing a label to consider for the plots (typically, the drug corresponding to the cv.res object)}

\item{modsel.criterion}{prediction performance criterion to select best model (balanced.accuracy.best or auc) (defaut = balanced.accuracy.best)}

\item{best.eps}{AUC tolerance considered to flag best model (defaut = 1)}
}
\description{
This function generates a figure (made of 4 individual plots) summarizing overall cross-validation results
}
\examples{
show_cv_overall()
}
