% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sparse_cor}
\alias{sparse_cor}
\title{sparse_cor}
\usage{
sparse_cor(x, y = NULL, return.cov = FALSE)
}
\arguments{
\item{x}{input matrix}

\item{y}{second input matrix (optional)}

\item{return.cov}{boolean flag to return covariance matrix in addition to correlation matrix (default : FALSE)}
}
\description{
This function computes a correlation matrix between the columns of a sparse matrix, or of a pair of sparse matrices.
}
\examples{
sparse_cor()
}
