% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{stratified_cv_folds}
\alias{stratified_cv_folds}
\title{stratified_cv_folds}
\usage{
stratified_cv_folds(y, n.folds)
}
\arguments{
\item{y}{reference labels}

\item{n.folds}{number of folds}
}
\description{
This function defines stratified cross-validation folds
}
\examples{
stratified_cv_folds()
}
